Micro optimize
