+++ src/Nethermind/Nethermind.Trie/Pruning/ITrieNodeResolver.cs
@@ -25,33 +25,4 @@ namespace Nethermind.Trie.Pruning
         /// <returns></returns>
         byte[]? LoadRlp(Keccak hash, ReadFlags flags = ReadFlags.None);
     }
-
-    public static class IBufferPoolExtensions
-    {
-        public static CappedArray<byte> SafeRentBuffer(this IBufferPool? pool, int size)
-        {
-            if (pool == null) return new CappedArray<byte>(new byte[size]);
-            CappedArray<byte> returnedBuffer = pool.RentBuffer(size);
-            if (returnedBuffer.Array == null)
-            {
-                return new CappedArray<byte>(new byte[size]);
-            }
-
-            return returnedBuffer;
-        }
-
-        public static void SafeReturnBuffer(this IBufferPool? pool, CappedArray<byte> buffer)
-        {
-            pool?.ReturnBuffer(buffer);
-        }
-    }
-
-    public interface IBufferPool
-    {
-        CappedArray<byte> RentBuffer(int size) => new CappedArray<byte>(new byte[size]);
-
-        void ReturnBuffer(CappedArray<byte> buffer)
-        {
-        }
-    }
 }
