+++ src/Nethermind/Nethermind.Trie/Pruning/TrieStore.cs
@@ -71,7 +71,7 @@ namespace Nethermind.Trie.Pruning
                     }
 
                     // we returning a copy to avoid multithreaded access
-                    trieNode = new TrieNode(NodeType.Unknown, hash, trieNode.FullRlp!.Value);
+                    trieNode = new TrieNode(NodeType.Unknown, hash, trieNode.FullRlp);
                     trieNode.ResolveNode(_trieStore);
                     trieNode.Keccak = hash;
